+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/ControlPlane.kt
@@ -224,7 +224,8 @@ class ControlPlane private constructor(
                 compositeDiscoveryServerCallbacks,
                 groupChangeWatcher,
                 executorGroup,
-                cachedProtoResourcesSerializer
+                cachedProtoResourcesSerializer,
+                properties.envoy.controlPlaneIdentifier
             )
         }